Manufacturer delete ehdot lisätty

+++ src/main/java/k23op1/op1_backend/web/ManufacturerController.java
@@ -1,5 +1,7 @@
 package k23op1.op1_backend.web;
 
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -8,6 +10,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+import org.springframework.web.servlet.view.RedirectView;
 
 import jakarta.validation.Valid;
 import k23op1.op1_backend.domain.Manufacturer;
@@ -43,11 +47,18 @@ public class ManufacturerController {
         return "listmanufacturers";
     }
 
-    // Valmistajan poisto
+    // Valmistajan poisto jos valmistajalla ei ole tuotteita, muuten virheilmoitus!
     @GetMapping("/deletemanufacturers/{id}")
-    public String deleteClothing(@PathVariable("id") Long manufacturerId, Model model) {
-        manufacturerRepository.deleteById(manufacturerId);
-        return "redirect:../listmanufacturers";
+    public RedirectView deleteClothing(@PathVariable("id") Long manufacturerId, RedirectAttributes attributes) {
+        Optional<Manufacturer> manufacturer = manufacturerRepository.findById(manufacturerId);
+        //muutetaan Optional normaaliksi
+        Manufacturer normalManufacturer = manufacturer.get();
+        if (normalManufacturer.getClothings().isEmpty()) {
+            manufacturerRepository.deleteById(manufacturerId);
+        }else {
+            attributes.addFlashAttribute("error", "Poistaminen epäonnistui, valmistajalla on tuotteita!");
+        }
+        return new RedirectView("/listmanufacturers");
     }
 
     // Valmistajan muokkaus
